Update change-case extension
- feat(change-case): add no-view commands for direct hotkey assignment
- feat(change-case): add no-view commands for direct hotkey assignment
