+++ extensions/change-case/src/utils.ts
@@ -1,0 +1,31 @@
+import { Clipboard, getSelectedText } from "@raycast/api";
+
+export class NoTextError extends Error {
+  constructor() {
+    super("No text");
+    Object.setPrototypeOf(this, NoTextError.prototype);
+  }
+}
+
+async function getSelection(): Promise<string> {
+  try {
+    return await getSelectedText();
+  } catch {
+    return "";
+  }
+}
+
+export async function readContent(preferredSource: string): Promise<string> {
+  const clipboard = (await Clipboard.readText()) ?? "";
+  const selected = await getSelection();
+
+  if (preferredSource === "clipboard") {
+    if (clipboard) return clipboard;
+    if (selected) return selected;
+  } else {
+    if (selected) return selected;
+    if (clipboard) return clipboard;
+  }
+
+  throw new NoTextError();
+}
